Added support for VS2017
Copied existing sln/project and let vs2017 perform auto-upgrade.

Copied HoseSky src into SampleFramework11 similar to
https://github.com/TheRealMJP/DeferredTexturing/commit/a8581f54edb79f05222c975916cb7a0e9ae85b81
to resolve link error.

Modified vs2013 and vs2015 projects to build from HoseSky src.

Also added pragma to MeshBaker.cpp to suppress a vs2013 warning.

+++ SampleFramework11/v1.02/HosekSky/ArHosekSkyModel.cpp
@@ -97,6 +97,8 @@ All instructions on how to use this code are in the accompanying header file.
 
 */
 
+#include "PCH.h"
+
 #include "ArHosekSkyModel.h"
 #include "ArHosekSkyModelData_Spectral.h"
 #include "ArHosekSkyModelData_CIEXYZ.h"
@@ -523,7 +525,7 @@ double arhosekskymodel_radiance(
         double                  wavelength
         )
 {
-    int low_wl = (wavelength - 320.0 ) / 40.0;
+    int low_wl = int((wavelength - 320.0 ) / 40.0);
 
     if ( low_wl < 0 || low_wl >= 11 )
         return 0.0f;
